did some changes to the matching game
